refactor(molsetup): stage 3 — extract rdkit_adapter, drop ABC
Move every RDKit-coupled operation off the RDKitMoleculeSetup class body
into free functions in ``meeko/molsetup/rdkit_adapter.py``:

  from_rdkit_mol, init_atom, init_bond, calculate_charges,
  get_charges_from_template, find_pattern, get_mol_name,
  get_smiles_and_order, perceive_rings,
  get_conformer_with_modified_positions,
  get_mol_with_modified_positions, get_num_mol_atoms,
  get_equivalent_atoms, restrain_to,
  add_dihedral_interaction, are_fourier_series_identical,
  get_symmetries_for_rmsd, has_implicit_hydrogens.

The ``MoleculeSetupExternalToolkit`` ABC mixin is deleted entirely — it
existed only to declare which methods RDKitMoleculeSetup needed to
provide, which is no longer meaningful once the implementations live in
their own module.

RDKitMoleculeSetup now inherits only from MoleculeSetup. Its methods are
preserved as one-line wrappers that delegate to rdkit_adapter, so
external callers (`setup.find_pattern(...)`, `RDKitMoleculeSetup.from_mol(...)`,
etc.) see no API change. Imports in setup.py shrank from 9 modules to 6.

All 92 tests pass (4 skipped).
